Added Semi-Automatic Pistol. Changed How Player Raycast Ignores Barrier.
Added a pistol to the game, limited how quickly it can fire, and made the animation speed static, as it's not going off continuously like an automatic gun. Still debating removing Melee weapons. Need to modify custom editor for weapon script to remove unnecesary details for semi-auto weapons. I also realized there is a layer to ignore raycasts, so i set the repair barrier interact wall and player barrier to ignore raycasts, and allowed ignore raycasts to be part of the interaction masks, since we're looking for specific game tags anyways.