Weapon Upgrade System Implementation & Integration With Point System
Implemented a workbench that upgrades your weapon for points. The weapon gets placed on the table and the upgrade bar begins. When it is complete the weapon upgrades. Need to revamp the inventory system to limit the number of weapons a player can carry, as well as to be able to remove and add weapons from a player's arsenal as needed when they are being upgraded/purchased.